+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 - Add new basic tests for the realty object model and mapper (#179)
 
 ### Changed
+- Skip the DB cleanup in the new functional tests (#249)
 - Require oelib >= 2.3.0 (#246)
 - Move the districts to within the cities with IRRE (#244)
 - Drop unneeded data scrubbing during import (#238)